+++ owncloudApp/src/main/java/com/owncloud/android/dependecyinjection/CommonModule.kt
@@ -21,12 +21,12 @@
 package com.owncloud.android.dependecyinjection
 
 import com.owncloud.android.presentation.manager.AvatarManager
+import com.owncloud.android.providers.AccountProvider
 import com.owncloud.android.providers.CameraUploadsHandlerProvider
 import com.owncloud.android.providers.ContextProvider
 import com.owncloud.android.providers.CoroutinesDispatcherProvider
 import com.owncloud.android.providers.LogsProvider
 import com.owncloud.android.providers.OCContextProvider
-import org.koin.android.ext.koin.androidApplication
 import org.koin.android.ext.koin.androidContext
 import org.koin.dsl.module
 
@@ -37,4 +37,5 @@ val commonModule = module {
     factory<ContextProvider> { OCContextProvider(androidContext()) }
     single { LogsProvider(get()) }
     single { CameraUploadsHandlerProvider(androidContext()) }
+    single { AccountProvider(androidContext()) }
 }