+++ owncloudApp/src/main/java/com/owncloud/android/presentation/viewmodels/settings/SettingsVideoUploadsViewModel.kt
@@ -24,7 +24,13 @@ import android.content.Intent
 import androidx.lifecycle.ViewModel
 import com.owncloud.android.data.preferences.datasources.SharedPreferencesProvider
 import com.owncloud.android.datamodel.OCFile
-import com.owncloud.android.db.PreferenceManager
+import com.owncloud.android.db.PreferenceManager.CameraUploadsConfiguration
+import com.owncloud.android.db.PreferenceManager.PREF__CAMERA_UPLOADS_DEFAULT_PATH
+import com.owncloud.android.db.PreferenceManager.PREF__CAMERA_VIDEO_UPLOADS_ACCOUNT_NAME
+import com.owncloud.android.db.PreferenceManager.PREF__CAMERA_VIDEO_UPLOADS_ENABLED
+import com.owncloud.android.db.PreferenceManager.PREF__CAMERA_VIDEO_UPLOADS_PATH
+import com.owncloud.android.db.PreferenceManager.PREF__CAMERA_VIDEO_UPLOADS_SOURCE
+import com.owncloud.android.providers.AccountProvider
 import com.owncloud.android.providers.CameraUploadsHandlerProvider
 import com.owncloud.android.ui.activity.LocalFolderPickerActivity
 import com.owncloud.android.ui.activity.UploadPathActivity
@@ -32,39 +38,65 @@ import java.io.File
 
 class SettingsVideoUploadsViewModel(
     private val preferencesProvider: SharedPreferencesProvider,
-    private val cameraUploadsHandlerProvider: CameraUploadsHandlerProvider
+    private val cameraUploadsHandlerProvider: CameraUploadsHandlerProvider,
+    private val accountProvider: AccountProvider
 ) : ViewModel() {
 
     fun isVideoUploadEnabled() =
-        preferencesProvider.getBoolean(PreferenceManager.PREF__CAMERA_VIDEO_UPLOADS_ENABLED, false)
+        preferencesProvider.getBoolean(PREF__CAMERA_VIDEO_UPLOADS_ENABLED, false)
 
-    fun setEnableVideoUpload(value: Boolean) =
-        preferencesProvider.putBoolean(PreferenceManager.PREF__CAMERA_VIDEO_UPLOADS_ENABLED, value)
+    fun setEnableVideoUpload(value: Boolean) {
+        preferencesProvider.putBoolean(PREF__CAMERA_VIDEO_UPLOADS_ENABLED, value)
+
+        if (value) {
+            // Use current account as default. It should never be null. If no accounts are attached, video uploads are hidden
+            accountProvider.getCurrentOwnCloudAccount()?.name?.let { name ->
+                preferencesProvider.putString(
+                    key = PREF__CAMERA_VIDEO_UPLOADS_ACCOUNT_NAME,
+                    value = name
+                )
+            }
+        } else {
+            // Reset fields after disabling the feature
+            preferencesProvider.removePreference(key = PREF__CAMERA_VIDEO_UPLOADS_PATH)
+            preferencesProvider.removePreference(key = PREF__CAMERA_VIDEO_UPLOADS_ACCOUNT_NAME)
+        }
+    }
 
     fun updateVideosLastSync() = cameraUploadsHandlerProvider.updateVideosLastSync(0)
 
+    fun getVideoUploadsAccount() = preferencesProvider.getString(
+        key = PREF__CAMERA_VIDEO_UPLOADS_ACCOUNT_NAME,
+        defaultValue = null
+    )
+
+    fun getLoggedAccountNames(): Array<String> = accountProvider.getLoggedAccounts().map { it.name }.toTypedArray()
+
     fun getVideoUploadsPath() = preferencesProvider.getString(
-        PreferenceManager.PREF__CAMERA_VIDEO_UPLOADS_PATH,
-        PreferenceManager.PREF__CAMERA_UPLOADS_DEFAULT_PATH
+        key = PREF__CAMERA_VIDEO_UPLOADS_PATH,
+        defaultValue = PREF__CAMERA_UPLOADS_DEFAULT_PATH
     )
 
     fun getVideoUploadsSourcePath() = preferencesProvider.getString(
-        PreferenceManager.PREF__CAMERA_VIDEO_UPLOADS_SOURCE,
-        PreferenceManager.CameraUploadsConfiguration.getDefaultSourcePath()
+        key = PREF__CAMERA_VIDEO_UPLOADS_SOURCE,
+        defaultValue = CameraUploadsConfiguration.getDefaultSourcePath()
     )
 
     fun handleSelectVideoUploadsPath(data: Intent?) {
         val folderToUpload = data?.getParcelableExtra<OCFile>(UploadPathActivity.EXTRA_FOLDER)
         folderToUpload?.remotePath?.let {
-            preferencesProvider.putString(PreferenceManager.PREF__CAMERA_VIDEO_UPLOADS_PATH, it)
+            preferencesProvider.putString(
+                key = PREF__CAMERA_VIDEO_UPLOADS_PATH,
+                value = it
+            )
         }
     }
 
     fun handleSelectVideoUploadsSourcePath(data: Intent?) {
         // If the source path has changed, update camera uploads last sync
         var previousSourcePath = preferencesProvider.getString(
-            PreferenceManager.PREF__CAMERA_VIDEO_UPLOADS_SOURCE,
-            PreferenceManager.CameraUploadsConfiguration.getDefaultSourcePath()
+            key = PREF__CAMERA_VIDEO_UPLOADS_SOURCE,
+            defaultValue = CameraUploadsConfiguration.getDefaultSourcePath()
         )
 
         previousSourcePath = previousSourcePath?.trimEnd(File.separatorChar)
@@ -75,7 +107,10 @@ class SettingsVideoUploadsViewModel(
         }
 
         data?.getStringExtra(LocalFolderPickerActivity.EXTRA_PATH)?.let {
-            preferencesProvider.putString(PreferenceManager.PREF__CAMERA_VIDEO_UPLOADS_SOURCE, it)
+            preferencesProvider.putString(
+                key = PREF__CAMERA_VIDEO_UPLOADS_SOURCE,
+                value = it
+            )
         }
     }