+++ owncloudData/src/main/java/com/owncloud/android/data/preferences/datasources/implementation/SharedPreferencesProviderImpl.kt
@@ -40,4 +40,6 @@ class SharedPreferencesProviderImpl(
 
     override fun putBoolean(key: String, value: Boolean) = editor.putBoolean(key, value).apply()
     override fun getBoolean(key: String, defaultValue: Boolean) = sharedPreferences.getBoolean(key, defaultValue)
+
+    override fun removePreference(key: String) = editor.remove(key).apply()
 }